# guardrail-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: Guardrail Gateway API
  version: 0.1.0
  description: Policy-enforced, auditable trust spine for agentic multi-enterprise systems
paths:
...
  /v1/explain/{decision_id}:
    get:
      summary: Get decision explanation
      parameters:
        - name: decision_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Decision explanation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionExplanation'
        '404':
          description: Decision not found
...
components:
  schemas:
...
    DecisionExplanation:
      type: object
      properties:
        decision_id: {type: string}
        agent: {type: object}
        resource: {type: object}
        allow: {type: boolean}
        policies_applied: {type: array, items: {type: string}}
        violations: {type: array, items: {type: object}}
        obligations: {type: array, items: {type: string}}
        explanation_text: {type: string}
        timestamp: {type: string, format: date-time}
        latency_ms: {type: integer}